Fix: update lib.rs imports order and fix generating of code
